# docs/api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "TravelGen 生成管线 API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/tasks/{task_id}": {
      "get": {
        "tags": [
          "项目分阶段"
        ],
        "summary": "Get Task Status",
        "operationId": "get_task_status_api_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/kb/search": {
      "get": {
        "tags": [
          "项目分阶段"
        ],
        "summary": "Search Kb",
        "operationId": "search_kb_api_kb_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Q"
            }
          },
          {
            "name": "top_k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 5,
              "title": "Top K"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/generate": {
      "post": {
        "summary": "Create Generate Task",
        "description": "提交生成任务（异步，202 返回 task_id）。",
        "operationId": "create_generate_task_api_v1_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}": {
      "get": {
        "summary": "Get Generate Task",
        "description": "轮询任务状态与阶段输出（前端 2s 间隔）。",
        "operationId": "get_generate_task_api_v1_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/kb/search": {
      "get": {
        "summary": "Search Kb",
        "description": "知识库检索（可选，供前端知识点预览）。",
        "operationId": "search_kb_api_v1_kb_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Q"
            }
          },
          {
            "name": "top_k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 5,
              "title": "Top K"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Asset": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "image"
          },
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "Asset"
      },
...
      "GenerateRequest": {
        "properties": {
          "city": {
            "type": "string",
            "title": "City"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "scene_type": {
            "type": "string",
            "title": "Scene Type"
          },
          "theme": {
            "type": "string",
            "title": "Theme"
          },
          "audience": {
            "type": "string",
            "title": "Audience",
            "default": "18-35岁年轻游客"
          },
          "style": {
            "type": "string",
            "title": "Style",
            "default": "大气唯美"
          },
          "duration_s": {
            "type": "integer",
            "maximum": 120.0,
            "minimum": 15.0,
            "title": "Duration S",
            "default": 60
          },
          "aspect_ratio": {
            "type": "string",
            "title": "Aspect Ratio",
            "default": "9:16"
          },
          "resolution": {
            "type": "string",
            "title": "Resolution",
            "default": "1080p"
          },
          "video_model": {
            "type": "string",
            "title": "Video Model",
            "default": "seedance-2.0-pro"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "assets": {
            "items": {
              "$ref": "#/components/schemas/Asset"
            },
            "type": "array",
            "title": "Assets",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "city",
          "location",
          "scene_type",
          "theme"
        ],
        "title": "GenerateRequest",
        "description": "输入契约（docs/API_Contract_MVP.md §2 / TravelGen_v1.md §六）。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}